move controller related functions/classes to controller_utils
