+++ sky/cli.py
@@ -60,12 +60,12 @@ from sky.backends import onprem_utils
 from sky.benchmark import benchmark_state
 from sky.benchmark import benchmark_utils
 from sky.clouds import service_catalog
-from sky.data import storage_utils
 from sky.skylet import constants
 from sky.skylet import job_lib
 from sky.usage import usage_lib
 from sky.utils import command_runner
 from sky.utils import common_utils
+from sky.utils import controller_utils
 from sky.utils import dag_utils
 from sky.utils import env_options
 from sky.utils import kubernetes_utils
@@ -1401,7 +1401,7 @@ def launch(
     """
     # NOTE(dev): Keep the docstring consistent between the Python API and CLI.
     env = _merge_env_vars(env_file, env)
-    backend_utils.check_cluster_name_not_reserved(
+    controller_utils.check_cluster_name_not_reserved(
         cluster, operation_str='Launching tasks on it')
     if backend_name is None:
         backend_name = backends.CloudVmRayBackend.NAME
@@ -1571,7 +1571,7 @@ def exec(
         raise ValueError('`ports` is not supported by `sky exec`.')
 
     env = _merge_env_vars(env_file, env)
-    backend_utils.check_cluster_name_not_reserved(
+    controller_utils.check_cluster_name_not_reserved(
         cluster, operation_str='Executing task on it')
     handle = global_user_state.get_handle_from_cluster_name(cluster)
     if handle is None:
@@ -1890,7 +1890,7 @@ def status(all: bool, refresh: bool, ip: bool, show_spot_jobs: bool,
         reserved_clusters = []
         for cluster_record in cluster_records:
             cluster_name = cluster_record['name']
-            controller = backend_utils.Controllers.check_cluster_name(
+            controller = controller_utils.Controllers.check_cluster_name(
                 cluster_name)
             if controller is not None:
                 reserved_clusters.append(cluster_record)
@@ -2024,7 +2024,8 @@ def cost_report(all: bool):  # pylint: disable=redefined-builtin
     reserved_clusters = dict()
     for cluster_record in cluster_records:
         cluster_name = cluster_record['name']
-        controller = backend_utils.Controllers.check_cluster_name(cluster_name)
+        controller = controller_utils.Controllers.check_cluster_name(
+            cluster_name)
         if controller is not None:
             controller_name = controller.value.name
             # to display most recent entry for each reserved cluster
@@ -2294,7 +2295,7 @@ def cancel(cluster: str, all: bool, jobs: List[int], yes: bool):  # pylint: disa
     try:
         core.cancel(cluster, all=all, job_ids=job_ids_to_cancel)
     except exceptions.NotSupportedError:
-        controller = backend_utils.Controllers.check_cluster_name(cluster)
+        controller = controller_utils.Controllers.check_cluster_name(cluster)
         assert controller is not None, cluster
         click.echo(controller.value.decline_cancel_hint)
         sys.exit(1)
@@ -2588,8 +2589,8 @@ def start(
         clusters = [
             cluster['name']
             for cluster in global_user_state.get_clusters()
-            if backend_utils.Controllers.check_cluster_name(cluster['name']) is
-            None
+            if controller_utils.Controllers.check_cluster_name(cluster['name'])
+            is None
         ]
 
     if not clusters:
@@ -2657,7 +2658,7 @@ def start(
     # Checks for reserved clusters (spot controller).
     reserved, non_reserved = [], []
     for name in to_start:
-        if backend_utils.Controllers.check_cluster_name(name) is not None:
+        if controller_utils.Controllers.check_cluster_name(name) is not None:
             reserved.append(name)
         else:
             non_reserved.append(name)
@@ -2775,7 +2776,8 @@ def _hint_or_raise_for_down_spot_controller(controller_name: str):
         click.echo('Managed spot controller has already been torn down.')
         return
 
-    controller = backend_utils.Controllers.check_cluster_name(controller_name)
+    controller = controller_utils.Controllers.check_cluster_name(
+        controller_name)
     assert controller is not None, controller_name
     if cluster_status == status_lib.ClusterStatus.INIT:
         with ux_utils.print_exception_no_traceback():
@@ -2827,7 +2829,8 @@ def _hint_or_raise_for_down_sky_serve_controller(controller_name: str):
         click.echo('Sky serve controller has already been torn down.')
         return
 
-    controller = backend_utils.Controllers.check_cluster_name(controller_name)
+    controller = controller_utils.Controllers.check_cluster_name(
+        controller_name)
     assert controller is not None, controller_name
     if cluster_status == status_lib.ClusterStatus.INIT:
         with ux_utils.print_exception_no_traceback():
@@ -2853,9 +2856,9 @@ def _hint_or_raise_for_down_sky_serve_controller(controller_name: str):
 
 
 _CONTROLLER_TO_HINT_OR_RAISE = {
-    backend_utils.Controllers.SPOT_CONTROLLER:
+    controller_utils.Controllers.SPOT_CONTROLLER:
         (_hint_or_raise_for_down_spot_controller),
-    backend_utils.Controllers.SKY_SERVE_CONTROLLER:
+    controller_utils.Controllers.SKY_SERVE_CONTROLLER:
         (_hint_or_raise_for_down_sky_serve_controller),
 }
 
@@ -2903,12 +2906,12 @@ def _down_or_stop_clusters(
     if len(names) > 0:
         reserved_clusters = [
             name for name in names
-            if backend_utils.Controllers.check_cluster_name(name) is not None
+            if controller_utils.Controllers.check_cluster_name(name) is not None
         ]
         reserved_clusters_str = ', '.join(map(repr, reserved_clusters))
         names = [
             name for name in _get_glob_clusters(names)
-            if backend_utils.Controllers.check_cluster_name(name) is None
+            if controller_utils.Controllers.check_cluster_name(name) is None
         ]
         if not down:
             local_clusters = onprem_utils.check_and_get_local_clusters()
@@ -2942,7 +2945,7 @@ def _down_or_stop_clusters(
                     f'{operation} reserved cluster(s) '
                     f'{reserved_clusters_str} is currently not supported.')
             else:
-                controller = backend_utils.Controllers.check_cluster_name(
+                controller = controller_utils.Controllers.check_cluster_name(
                     reserved_cluster)
                 assert controller is not None
                 hint_or_raise = _CONTROLLER_TO_HINT_OR_RAISE[controller]
@@ -2967,8 +2970,10 @@ def _down_or_stop_clusters(
         # Otherwise, it would be very easy to accidentally delete a reserved
         # cluster.
         names = [
-            record['name'] for record in all_clusters if
-            backend_utils.Controllers.check_cluster_name(record['name']) is None
+            record['name']
+            for record in all_clusters
+            if controller_utils.Controllers.check_cluster_name(record['name'])
+            is None
         ]
 
     clusters = []
@@ -3580,7 +3585,7 @@ def storage():
 def storage_ls(all: bool):
     """List storage objects managed by SkyPilot."""
     storages = sky.storage_ls()
-    storage_table = storage_utils.format_storage_table(storages, show_all=all)
+    storage_table = status_utils.format_storage_table(storages, show_all=all)
     click.echo(storage_table)
 
 
@@ -4018,8 +4023,8 @@ def spot_cancel(name: Optional[str], job_ids: Tuple[int], all: bool, yes: bool):
       # Cancel managed spot jobs with IDs 1, 2, 3
       $ sky spot cancel 1 2 3
     """
-    _, handle = backend_utils.is_controller_up(
-        controller_type=backend_utils.Controllers.SPOT_CONTROLLER,
+    _, handle = controller_utils.is_controller_up(
+        controller_type=controller_utils.Controllers.SPOT_CONTROLLER,
         stopped_message='All managed spot jobs should have finished.')
     if handle is None:
         # Hint messages already printed by the call above.
@@ -4103,8 +4108,8 @@ def spot_dashboard(port: Optional[int]):
     hint = (
         'Dashboard is not available if spot controller is not up. Run a spot '
         'job first.')
-    _, handle = backend_utils.is_controller_up(
-        controller_type=backend_utils.Controllers.SPOT_CONTROLLER,
+    _, handle = controller_utils.is_controller_up(
+        controller_type=controller_utils.Controllers.SPOT_CONTROLLER,
         stopped_message=hint,
         non_existent_message=hint)
     if handle is None:
@@ -4182,7 +4187,7 @@ def serve_up(
         sky serve up service.yaml
     """
     if service_name is None:
-        service_name = backend_utils.generate_service_name()
+        service_name = serve_lib.generate_service_name()
 
     is_yaml, _ = _check_yaml(''.join(service_yaml))
     if not is_yaml:
@@ -4378,8 +4383,8 @@ def serve_down(service_names: List[str], all: bool, yes: bool):
             'Can only specify one of SERVICE_NAMES or --all. '
             f'Provided {argument_str!r}.')
 
-    _, handle = backend_utils.is_controller_up(
-        controller_type=backend_utils.Controllers.SKY_SERVE_CONTROLLER,
+    _, handle = controller_utils.is_controller_up(
+        controller_type=controller_utils.Controllers.SKY_SERVE_CONTROLLER,
         stopped_message='All services should have been terminated.')
     if handle is None:
         # Hint messages already printed by the call above.
