+++ sky/spot/spot_utils.py
@@ -36,7 +36,8 @@ logger = sky_logging.init_logger(__name__)
 
 # Add user hash so that two users don't have the same controller VM on
 # shared-account clouds such as GCP.
-SPOT_CONTROLLER_NAME = f'sky-spot-controller-{common_utils.get_user_hash()}'
+SPOT_CONTROLLER_NAME: str = (
+    f'sky-spot-controller-{common_utils.get_user_hash()}')
 SIGNAL_FILE_PREFIX = '/tmp/sky_spot_controller_signal_{}'
 # Controller checks its job's status every this many seconds.
 JOB_STATUS_CHECK_GAP_SECONDS = 20
